Created class hierarchy
